Moved SelfDestructingResource from utils to io package
